+++ src/ngx_http_lua_req_body.c
@@ -558,7 +558,7 @@ ngx_http_lua_ngx_req_init_body(lua_State *L)
 
         /* avoid allocating an unnecessary large buffer */
         if (size > (size_t) r->headers_in.content_length_n) {
-            size = r->headers_in.content_length_n;
+            size = (size_t) r->headers_in.content_length_n;
         }
     }
 
@@ -738,7 +738,7 @@ ngx_http_lua_ngx_req_body_finish(lua_State *L)
         return luaL_error(L, "out of memory");
     }
 
-    size = r->headers_in.content_length_n;
+    size = (size_t) r->headers_in.content_length_n;
 
     value.len = ngx_sprintf(value.data, "%uz", size) - value.data;
     value.data[value.len] = '\0';
